# Conflicts:
#	src/platz1de/EasyEdit/selection/DynamicBlockListSelection.php
#	src/platz1de/EasyEdit/task/selection/PasteTask.php

+++ src/platz1de/EasyEdit/utils/MixedUtils.php
@@ -7,13 +7,13 @@ use pocketmine\Server;
 class MixedUtils
 {
 	/**
-	 * @param string $int
-	 * @param bool   $data
+	 * @param int  $int
+	 * @param bool $data
 	 * @return string
 	 */
-	public static function humanReadable(string $int, bool $data = false): string
+	public static function humanReadable(int $int, bool $data = false): string
 	{
-		$factor = $data ? floor(log($int, 1024)) : floor((strlen($int) - 1) / 3);
+		$factor = $data ? floor(log($int, 1024)) : floor((strlen((string) $int) - 1) / 3);
 		$string = $data ? number_format($int / 1024 ** $factor) : number_format($int);
 		$unit = $data ? ["B", "KB", "MB", "GB", "TB"] : ["", "K", "M", "G", "T"];
 		return substr($string, 0, strpos($string, ",") < 3 ? 4 : 3) . $unit[$factor];
@@ -28,8 +28,11 @@ class MixedUtils
 			return;
 		}
 
-		$files = array_diff(scandir($dir), ['.', '..']);
-		foreach ($files as $file) {
+		$files = scandir($dir);
+		if ($files === false) {
+			return;
+		}
+		foreach (array_diff($files, ['.', '..']) as $file) {
 			is_dir($dir . DIRECTORY_SEPARATOR . $file) ? self::deleteDir($dir . DIRECTORY_SEPARATOR . $file) : unlink($dir . DIRECTORY_SEPARATOR . $file);
 		}
 		rmdir($dir);
@@ -41,7 +44,7 @@ class MixedUtils
 	 */
 	public static function setAutoSave(int $cooldown): int
 	{
-		return (function () use ($cooldown) {
+		return (function () use ($cooldown): int {
 			$previous = $this->autoSaveTicks;
 			$this->autoSaveTicks = $cooldown;
 			return $previous;
